Add rs-query-provider export and refactor user store
