+++ libs/react-query/src/index.ts
@@ -10,4 +10,4 @@ export * from './role';
 export * from './user';
 export * from './utils';
 
-export * from './providers/rs-query-provider';
+export * from './providers';
