+++ libs/react-query/src/user/user.store.ts
@@ -4,12 +4,14 @@ import { zustandStore } from '../utils/zustand.store';
 type UserState = {
   user: any;
   totalUser: number;
+  users: any;
 };
 
 type UserStateAction = {
   setUser: (user: any) => void;
   clearUser: () => void;
   setTotalUser: (totalUser: number) => void;
+  setUsers: (users: any) => void;
 };
 
 type UserStore = UserState & UserStateAction;
@@ -24,6 +26,8 @@ export const useUserStore = zustandStore<UserStore>(
       set({
         totalUser,
       }),
+    users: [],
+    setUsers: (users) => set({ users }),
   }),
   {
     devtoolsEnabled: true,
